test_computer_vision/requirements.txt
torch
torchvision
pillow
scikit-learn
tabulate